Factored out `pyenki_viewer`
