+++ python/pyenki/pyenki.pyi
@@ -1,7 +1,9 @@
 import collections.abc
+import typing
+import enum
 import numpy
 import numpy.typing
-import typing
+
 from typing import Annotated, ClassVar, TypeAlias, overload
 
 Vector: TypeAlias = Annotated[numpy.typing.NDArray[numpy.float64], '[2, 1]']
@@ -11,7 +13,9 @@ Array2D: TypeAlias = Annotated[numpy.typing.NDArray[numpy.float64], '[n, m]']
 IntArray1D: TypeAlias = Annotated[numpy.typing.NDArray[numpy.int64], '[n, m]']
 Image: TypeAlias = Annotated[numpy.typing.NDArray[numpy.uint8], '[n, m, 3]']
 Polygon: TypeAlias = collections.abc.Sequence[Vector]
-Part: TypeAlias = tuple[Polygon, typing.SupportsFloat] | tuple[Polygon, typing.SupportsFloat, collections.abc.Sequence[Color]]
+ARGBImage: TypeAlias = Annotated[numpy.typing.NDArray[numpy.uint8], '[n, m, 4]']
+ARGBImageLike: TypeAlias = Annotated[numpy.typing.ArrayLike, numpy.uint8, '[n, m, 4]']
+
 
 class Color:
     black: ClassVar[Color] = ...  # read-only
@@ -98,6 +102,22 @@ class Marxbot(DifferentialWheeled):
     def scanner_image(self) -> Array2D: ...
 
 class PhysicalObject:
+
+    class Part:
+        @overload
+        def __init__(self, l1: typing.SupportsFloat, l2: typing.SupportsFloat, height: typing.SupportsFloat) -> None: ...
+        @overload
+        def __init__(self, shape: collections.abc.Sequence[VectorLike], height: typing.SupportsFloat, textures: collections.abc.Sequence[collections.abc.Sequence[Color]] = ...) -> None: ...
+        def __eq__(self, arg0: object) -> bool: ...
+        def __hash__(self) -> int: ...
+        def __ne__(self, arg0: object) -> bool: ...
+        @property
+        def height(self) -> float: ...
+        @property
+        def shape(self) -> list[Vector]: ...
+        @property
+        def textures(self) -> list[list[Color]]: ...
+
     angle: float
     angular_speed: float
     collision_callback: collections.abc.Callable[[PhysicalObject, PhysicalObject], None]
@@ -106,6 +126,14 @@ class PhysicalObject:
     control_step_callback: collections.abc.Callable[[PhysicalObject, typing.SupportsFloat], None]
     dry_friction_coefficient: float
     name: str
+    @overload
+    def __init__(self, parts: collections.abc.Sequence[PhysicalObject.Part], mass: typing.SupportsFloat, color: Color = ...) -> None: ...
+    @overload
+    def __init__(self, radius: typing.SupportsFloat, height: typing.SupportsFloat, mass: typing.SupportsFloat, color: Color = ...) -> None: ...
+    @overload
+    def __init__(self, shape: collections.abc.Sequence[Vector], height: typing.SupportsFloat, mass: typing.SupportsFloat, color: Color = ..., face_colors: collections.abc.Sequence[Color] = ...) -> None: ...
+    @overload
+    def __init__(self, l1: typing.SupportsFloat, l2: typing.SupportsFloat, height: typing.SupportsFloat, mass: typing.SupportsFloat, color: Color = ...) -> None: ...
     @property
     def position(self) -> Vector: ...
     @position.setter
@@ -135,6 +163,8 @@ class PhysicalObject:
     def uid(self) -> int: ...
     @property
     def world(self) -> World: ...
+    @property
+    def parts(self) -> list[PhysicalObject.Part]: ...
 
 class Robot(PhysicalObject):
     pass
@@ -175,24 +205,40 @@ class Thymio2(DifferentialWheeled):
     def prox_values(self) -> Array1D: ...
     @property
     def prox_values_i(self) -> IntArray1D: ...
+    @property
+    def led_colors(self) -> numpy.typing.NDArray[numpy.float64]: ...
 
 class World:
+    class WallsType(enum.Enum):
+        SQUARE = ...
+        CIRCULAR = ...
+        NONE = ...
+
+    class GroundTexture:
+        @overload
+        def __init__(self) -> None: ...
+        @overload
+        def __init__(self, data: ARGBImageLike) -> None: ...
+        @property
+        def height(self) -> int: ...
+        @property
+        def width(self) -> int: ...
+        def __buffer__(self, flags: int) -> memoryview: ...
+        def __release_buffer__(self, view: memoryview) -> None: ...
+
     control_step_callback: collections.abc.Callable[[World, typing.SupportsFloat], None]
     random_generator: numpy.random.Generator
     random_seed: int
     @overload
     def __init__(self, seed: typing.SupportsInt = ...) -> None: ...
     @overload
-    def __init__(self, width: typing.SupportsFloat, height: typing.SupportsFloat, walls_color: Color = ..., seed: typing.SupportsInt = ...) -> None: ...
+    def __init__(self, width: typing.SupportsFloat, height: typing.SupportsFloat, walls_color: Color = ..., ground_texture: World.GroundTexture = ..., seed: typing.SupportsInt = ...) -> None: ...
     @overload
-    def __init__(self, radius: typing.SupportsFloat, walls_color: Color = ..., seed: typing.SupportsInt = ...) -> None: ...
+    def __init__(self, radius: typing.SupportsFloat, walls_color: Color = ..., ground_texture: World.GroundTexture = ..., seed: typing.SupportsInt = ...) -> None: ...
     def add_object(self, object: PhysicalObject) -> None: ...
     def copy_random_generator(self, world: World) -> None: ...
     def remove_object(self, object: PhysicalObject) -> None: ...
-    def render(self, camera_reset: bool = ..., camera_position: VectorLike = ..., camera_altitude: typing.SupportsFloat = ..., camera_yaw: typing.SupportsFloat = ..., camera_pitch: typing.SupportsFloat = ..., camera_is_ortho: bool = ..., walls_height: typing.SupportsFloat = ..., width: typing.SupportsFloat = ..., height: typing.SupportsFloat = ...) -> Image: ...
     def run(self, steps: typing.SupportsInt = ..., time_step: typing.SupportsFloat = ..., physics_oversampling: typing.SupportsInt = ..., termination: collections.abc.Callable[[World], bool] | None = ..., callback: collections.abc.Callable[[World], None] | None = ...) -> None: ...
-    def run_in_viewer(self, fps: typing.SupportsFloat = ..., time_step: typing.SupportsFloat = ..., factor: typing.SupportsFloat = ..., helpers: bool = ..., camera_reset: bool = ..., camera_position: VectorLike = ..., camera_altitude: typing.SupportsFloat = ..., camera_yaw: typing.SupportsFloat = ..., camera_pitch: typing.SupportsFloat = ..., camera_is_ortho: bool = ..., walls_height: typing.SupportsFloat = ..., duration: typing.SupportsFloat = ...) -> None: ...
-    def save_image(self, path: str, camera_reset: bool = ..., camera_position: VectorLike = ..., camera_altitude: typing.SupportsFloat = ..., camera_yaw: typing.SupportsFloat = ..., camera_pitch: typing.SupportsFloat = ..., camera_is_ortho: bool = ..., walls_height: typing.SupportsFloat = ..., width: typing.SupportsFloat = ..., height: typing.SupportsFloat = ...) -> None: ...
     def step(self, time_step: typing.SupportsFloat, physics_oversampling: typing.SupportsInt = ...) -> None: ...
     @property
     def objects(self) -> set[PhysicalObject]: ...
@@ -200,40 +246,18 @@ class World:
     def robots(self) -> list[Robot]: ...
     @property
     def static_objects(self) -> list[PhysicalObject]: ...
-
-class WorldView:
-    camera_altitude: float
-    camera_is_ortho: bool
-    camera_pitch: float
-    camera_pose: tuple[float, float, float, float]
     @property
-    def camera_position(self) -> Vector: ...
-    @camera_position.setter
-    def camera_position(self, value: VectorLike) -> None: ...
-    camera_yaw: float
-    helpers: bool
-    tracking: bool
-    world: World
-    def __init__(self, world: World = ..., fps: typing.SupportsFloat = ..., update_world: bool = ..., time_step: typing.SupportsFloat = ..., factor: typing.SupportsFloat = ..., helpers: bool = ..., camera_reset: bool = ..., camera_position: VectorLike = ..., camera_altitude: typing.SupportsFloat = ..., camera_yaw: typing.SupportsFloat = ..., camera_pitch: typing.SupportsFloat = ..., camera_is_ortho: bool = ..., walls_height: typing.SupportsFloat = ...) -> None: ...
-    def hide(self) -> None: ...
-    def move_camera(self, target_position: VectorLike, target_altitude: typing.SupportsFloat = ..., target_distance: typing.SupportsFloat = ..., yaw: typing.SupportsFloat | None = ..., pitch: typing.SupportsFloat | None = ...) -> None: ...
-    def point_camera(self, target_position: VectorLike, target_altitude: typing.SupportsFloat = ..., position: VectorLike | None = ..., altitude: typing.SupportsFloat | None = ...) -> None: ...
-    def reset_camera(self) -> None: ...
-    def save_image(self, arg0: str) -> None: ...
-    def show(self) -> None: ...
-    def start_updating_world(self, time_step: typing.SupportsFloat = ..., factor: typing.SupportsFloat = ...) -> None: ...
-    def stop_updating_world(self) -> None: ...
+    def radius(self) -> float: ...
     @property
-    def image(self) -> Image: ...
+    def width(self) -> float: ...
     @property
-    def walls_height(self) -> float: ...
+    def height(self) -> float: ...
     @property
-    def widget(self) -> object: ...
-
-def CircularObject(radius: typing.SupportsFloat, height: typing.SupportsFloat, mass: typing.SupportsFloat, color: Color = ...) -> PhysicalObject: ...
-def CompositeObject(parts: list[Part], mass: typing.SupportsFloat, color: Color = ...) -> PhysicalObject: ...
-def ConvexObject(base: collections.abc.Sequence[Vector], height: typing.SupportsFloat, mass: typing.SupportsFloat, color: Color = ..., face_colors: collections.abc.Sequence[Color] = ...) -> PhysicalObject: ...
-def RectangularObject(l1: typing.SupportsFloat, l2: typing.SupportsFloat, height: typing.SupportsFloat, mass: typing.SupportsFloat, color: Color = ...) -> PhysicalObject: ...
-def cleanup_ui() -> None: ...
-def init_ui() -> None: ...
-def run_ui(duration: typing.SupportsFloat = ...) -> None: ...
+    def walls_type(self) -> World.WallsType: ...
+    @property
+    def walls_color(self) -> Color: ...
+    @property
+    def ground_texture(self) -> World.GroundTexture: ...
+    @property
+    def has_ground_texture(self) -> bool: ...
+    def get_ground_color(self, position: VectorLike) -> Color: ...
