+++ python/pyenki/pyenki_viewer.pyi
@@ -1,0 +1,59 @@
+import numpy
+import numpy.typing
+import typing
+from typing import Annotated, TypeAlias
+from .pyenki import World
+
+Vector: TypeAlias = Annotated[numpy.typing.NDArray[numpy.float64], '[2, 1]']
+VectorLike: TypeAlias = Annotated[numpy.typing.ArrayLike, numpy.float64, '[2, 1]']
+Image: TypeAlias = Annotated[numpy.typing.NDArray[numpy.uint8], '[n, m, 3]']
+
+def cleanup() -> None: ...
+def init(share: bool = ...) -> None: ...
+def run(duration: typing.SupportsFloat = ...) -> None: ...
+def render(world: World, camera_reset: bool = ..., camera_position: VectorLike = ..., camera_altitude: typing.SupportsFloat = ..., camera_yaw: typing.SupportsFloat = ..., camera_pitch: typing.SupportsFloat = ..., camera_is_ortho: bool = ..., walls_height: typing.SupportsFloat = ..., width: typing.SupportsInt = ..., height: typing.SupportsInt = ...) -> Image: ...
+def run_in_viewer(world: World, fps: typing.SupportsFloat = ..., time_step: typing.SupportsFloat = ..., factor: typing.SupportsFloat = ..., helpers: bool = ..., camera_reset: bool = ..., camera_position: VectorLike = ..., camera_altitude: typing.SupportsFloat = ..., camera_yaw: typing.SupportsFloat = ..., camera_pitch: typing.SupportsFloat = ..., camera_is_ortho: bool = ..., walls_height: typing.SupportsFloat = ..., duration: typing.SupportsFloat = ...) -> None: ...
+def save_image(world: World, path: str, camera_reset: bool = ..., camera_position: VectorLike = ..., camera_altitude: typing.SupportsFloat = ..., camera_yaw: typing.SupportsFloat = ..., camera_pitch: typing.SupportsFloat = ..., camera_is_ortho: bool = ..., walls_height: typing.SupportsFloat = ..., width: typing.SupportsInt = ..., height: typing.SupportsInt = ...) -> None: ...
+
+class WorldView:
+    @property
+    def camera_altitude(self) -> float: ...
+    @camera_altitude.setter
+    def camera_altitude(self, value: typing.SupportsFloat) -> None: ...
+    camera_is_ortho: bool
+    @property
+    def camera_yaw(self) -> float: ...
+    @camera_yaw.setter
+    def camera_yaw(self, value: typing.SupportsFloat) -> None: ...
+    @property
+    def camera_pitch(self) -> float: ...
+    @camera_pitch.setter
+    def camera_pitch(self, value: typing.SupportsFloat) -> None: ...
+    @property
+    def camera_pose(self) -> tuple[Vector, float, float, float]: ...
+    @camera_pose.setter
+    def camera_pose(self, value: tuple[VectorLike, typing.SupportsFloat, typing.SupportsFloat, typing.SupportsFloat]) -> None: ...
+    @property
+    def camera_position(self) -> Vector: ...
+    @camera_position.setter
+    def camera_position(self, value: VectorLike) -> None: ...
+    helpers: bool
+    tracking: bool
+    world: World
+    def __init__(self, world: World = ..., fps: typing.SupportsFloat = ..., update_world: bool = ..., time_step: typing.SupportsFloat = ..., factor: typing.SupportsFloat = ..., helpers: bool = ..., camera_reset: bool = ..., camera_position: VectorLike = ..., camera_altitude: typing.SupportsFloat = ..., camera_yaw: typing.SupportsFloat = ..., camera_pitch: typing.SupportsFloat = ..., camera_is_ortho: bool = ..., walls_height: typing.SupportsFloat = ...) -> None: ...
+    def hide(self) -> None: ...
+    def move_camera(self, target_position: VectorLike, target_altitude: typing.SupportsFloat = ..., target_distance: typing.SupportsFloat = ..., yaw: typing.SupportsFloat | None = ..., pitch: typing.SupportsFloat | None = ...) -> None: ...
+    def point_camera(self, target_position: VectorLike, target_altitude: typing.SupportsFloat = ..., position: VectorLike | None = ..., altitude: typing.SupportsFloat | None = ...) -> None: ...
+    def reset_camera(self) -> None: ...
+    def save_image(self, arg0: str) -> None: ...
+    def show(self) -> None: ...
+    def start_updating_world(self, time_step: typing.SupportsFloat = ..., factor: typing.SupportsFloat = ...) -> None: ...
+    def stop_updating_world(self) -> None: ...
+    @property
+    def image(self) -> Image: ...
+    @property
+    def walls_height(self) -> float: ...
+    @property
+    def pyside_widget(self) -> object: ...
+    @property
+    def pyqt_widget(self) -> object: ...
